# app/docs/swagger.yaml
definitions:
...
  config.UpdateResult:
    description: Outcome when applying changes to the configuration file.
    properties:
      applied:
        description: Parameters applied immediately
        items:
          type: string
        type: array
      rejected:
        additionalProperties:
          type: string
        description: Parameters rejected with reasons
        type: object
      requires_restart:
        description: Parameters that require a restart
        items:
          type: string
        type: array
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.MultiGetResponse:
    additionalProperties:
      $ref: '#/definitions/model.Document'
    description: Map of keys to documents or null for missing entries.
    type: object
...
  handlers.incrementRequest:
    properties:
      delta:
        description: The amount to increment (positive) or decrement (negative) the
          counter.
        type: integer
    type: object
  handlers.incrementResponse:
    properties:
      new:
        description: New value
        type: integer
      old:
        description: Previous value
        type: integer
    type: object
...
  handlers.multiGetRequest:
    properties:
      keys:
        items:
          type: string
        type: array
    type: object
...
  model.Document:
    properties:
      key:
        description: Logical key
        type: string
      meta:
        allOf:
        - $ref: '#/definitions/model.Metadata'
        description: Metadata for versioning, expiration, etc.
      value:
        description: 'Content: string, int, []string, map[string]any, etc.'
    type: object
  model.Metadata:
    properties:
      created_at:
        description: When document was created
        type: string
      expiration:
        description: TTL as Unix timestamp (0 = never)
        type: integer
      rev:
        description: Revision ID for conflict resolution
        type: string
      type:
        description: Document type (json, counter, list, set)
        type: string
      updated_at:
        description: When document was last updated
        type: string
    type: object
info:
  contact: {}
paths:
...
  /config/update:
    put:
      consumes:
      - application/json
      description: Updates system configuration parameters. Some changes are applied
        live, while others require a restart.
      parameters:
      - description: Configuration parameters to update
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/config.UpdateResult'
        "400":
          description: Invalid JSON body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update configuration
      tags:
      - config
  /documents:
    delete:
      description: Deletes a document by key within a specified column family.
      parameters:
      - description: Document key to delete
        in: query
        name: key
        required: true
        type: string
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Document successfully deleted
        "400":
          description: Missing or invalid parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Document or column family not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete a document
      tags:
      - documents
    get:
      description: Retrieves a document by key, including its value and metadata.
      parameters:
      - description: Document key
        in: query
        name: key
        required: true
        type: string
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      - description: Optional RocksDB fill cache read option
        in: query
        name: fill_cache
        type: boolean
      - description: Optional RocksDB read tier (e.g. 'all', 'cache-only')
        in: query
        name: read_tier
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Document'
        "400":
          description: Missing or invalid key
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Retrieve a document
      tags:
      - documents
    post:
      consumes:
      - application/json
      description: Stores a document in the specified column family. Supports optimistic
        concurrency via CAS.
      parameters:
      - description: Document key
        in: query
        name: key
        required: true
        type: string
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      - description: Document type (e.g. 'json', 'counter', 'list')
        in: query
        name: type
        type: string
      - description: CAS (revision) for concurrency control
        in: query
        name: cas
        type: string
      - description: Document value (JSON-encoded)
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Document'
        "400":
          description: Invalid input or missing value
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Column family not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: CAS mismatch
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Store or update a document
      tags:
      - documents
  /documents/bulk/get:
    post:
      consumes:
      - application/json
      description: Retrieves multiple documents with metadata by key. Missing keys
        will be returned with null values.
      parameters:
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      - description: List of keys to retrieve
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.multiGetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MultiGetResponse'
        "400":
          description: Invalid JSON or missing key list
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Bulk document fetch
      tags:
      - documents
...
  /documents/counters/delta:
    post:
      consumes:
      - application/json
      description: Increments or decrements a counter document by a given integer
        value.
      parameters:
      - description: Document key
        in: query
        name: key
        required: true
        type: string
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      - description: Delta value for increment or decrement
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.incrementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.incrementResponse'
        "400":
          description: Invalid parameters or JSON body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Modify counter
      tags:
      - counters
...
  /documents/keys:
    get:
      description: Returns a list of keys within the specified column family, optionally
        filtered by prefix and pagination options.
      parameters:
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      - description: Only return keys with this prefix
        in: query
        name: prefix
        type: string
      - description: Return keys after this key (for pagination)
        in: query
        name: start_after
        type: string
      - description: 'Maximum number of keys to return (default: 100)'
        in: query
        name: limit
        type: integer
      - description: Whether to fill RocksDB read cache
        in: query
        name: fill_cache
        type: boolean
      - description: RocksDB read tier (e.g. 'all', 'cache-only')
        in: query
        name: read_tier
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List document keys
      tags:
      - keys
  /documents/list:
    get:
      description: Returns a map of documents filtered by optional prefix and paginated
        using start_after and limit.
      parameters:
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      - description: Filter documents whose keys begin with this prefix
        in: query
        name: prefix
        type: string
      - description: Skip documents until this key (exclusive)
        in: query
        name: start_after
        type: string
      - description: 'Maximum number of documents to return (default: 100)'
        in: query
        name: limit
        type: integer
      - description: Whether to fill RocksDB read cache
        in: query
        name: fill_cache
        type: boolean
      - description: RocksDB read tier (e.g., 'all', 'cache-only')
        in: query
        name: read_tier
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/model.Document'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List documents
      tags:
      - documents
...
swagger: "2.0"